// src/api/webrtc-server/model/messageBodyRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (http://openapi-generator.tech).
 * http://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Server } from './server';
import { WebClient } from './webClient';


export interface MessageBodyRequest { 
    requestId?: string;
    readonly type: MessageBodyRequest.TypeEnum;
    server?: Server;
    clients?: Array<WebClient>;
}
export namespace MessageBodyRequest {
    export const TypeEnum = {
        MessageBodyRequest: 'MessageBodyRequest'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}